refactor: User Settings
