+++ mail/client/doctype/account_settings/test_account_settings.py
@@ -11,9 +11,9 @@ EXTRA_TEST_RECORD_DEPENDENCIES = []  # eg. ["User"]
 IGNORE_TEST_RECORD_DEPENDENCIES = []  # eg. ["User"]
 
 
-class IntegrationTestAccountState(IntegrationTestCase):
+class IntegrationTestAccountSettings(IntegrationTestCase):
 	"""
-	Integration tests for AccountState.
+	Integration tests for AccountSettings.
 	Use this class for testing interactions between multiple components.
 	"""
