--- mail/client/doctype/account_state/account_state.js
@@ -1,8 +1,0 @@
-// Copyright (c) 2026, Frappe Technologies Pvt. Ltd. and contributors
-// For license information, please see license.txt
-
-// frappe.ui.form.on("Account State", {
-// 	refresh(frm) {
-
-// 	},
-// });
