+++ mail/client/doctype/user_settings/user_settings.json
@@ -11,14 +11,6 @@
   "column_break_uxkb",
   "username",
   "app_password",
-  "cache_section",
-  "has_cached_jmap_connection",
-  "has_cached_jmap_identities",
-  "has_cached_jmap_mailboxes",
-  "column_break_uuod",
-  "total_cached_blobs",
-  "total_cached_mail_messages",
-  "total_cached_contact_cards",
   "outgoing_section",
   "default_outgoing_email",
   "column_break_scmb",
@@ -206,71 +198,6 @@
    "search_index": 1
   },
   {
-   "collapsible": 1,
-   "depends_on": "eval: doc.username",
-   "fieldname": "cache_section",
-   "fieldtype": "Section Break",
-   "label": "Cache"
-  },
-  {
-   "default": "0",
-   "fieldname": "has_cached_jmap_connection",
-   "fieldtype": "Check",
-   "is_virtual": 1,
-   "label": "Has Cached JMAP Connection",
-   "no_copy": 1,
-   "read_only": 1
-  },
-  {
-   "default": "0",
-   "fieldname": "has_cached_jmap_identities",
-   "fieldtype": "Check",
-   "is_virtual": 1,
-   "label": "Has Cached JMAP Identities",
-   "no_copy": 1,
-   "read_only": 1
-  },
-  {
-   "default": "0",
-   "fieldname": "has_cached_jmap_mailboxes",
-   "fieldtype": "Check",
-   "is_virtual": 1,
-   "label": "Has Cached JMAP Mailboxes",
-   "no_copy": 1,
-   "read_only": 1
-  },
-  {
-   "fieldname": "column_break_uuod",
-   "fieldtype": "Column Break"
-  },
-  {
-   "default": "0",
-   "fieldname": "total_cached_mail_messages",
-   "fieldtype": "Int",
-   "is_virtual": 1,
-   "label": "Total Cached Mail Messages",
-   "no_copy": 1,
-   "read_only": 1
-  },
-  {
-   "default": "0",
-   "fieldname": "total_cached_contact_cards",
-   "fieldtype": "Int",
-   "is_virtual": 1,
-   "label": "Total Cached Contact Cards",
-   "no_copy": 1,
-   "read_only": 1
-  },
-  {
-   "default": "0",
-   "fieldname": "total_cached_blobs",
-   "fieldtype": "Int",
-   "is_virtual": 1,
-   "label": "Total Cached Blobs",
-   "no_copy": 1,
-   "read_only": 1
-  },
-  {
    "fieldname": "contacts_section",
    "fieldtype": "Section Break",
    "label": "Contacts"
@@ -304,7 +231,7 @@
  "grid_page_length": 50,
  "index_web_pages_for_search": 1,
  "links": [],
- "modified": "2026-04-15 16:07:15.843640",
+ "modified": "2026-04-16 11:30:00.537911",
  "modified_by": "Administrator",
  "module": "Client",
  "name": "User Settings",
